finish settings
