+++ FanControl/build.gradle.kts
@@ -12,8 +12,8 @@ repositories {
     google()
     mavenCentral()
     maven("https://maven.pkg.jetbrains.space/public/p/compose/dev")
-    //maven("https://jitpack.io")
-    mavenLocal()
+    maven("https://jitpack.io")
+    //mavenLocal()
 }
 
 kotlin {
@@ -33,8 +33,8 @@ kotlin {
 
                 implementation("org.jetbrains.compose.material3:material3-desktop:${extra["compose.version"] as String}")
                 implementation("org.json:json:20220924")
-                //implementation("com.github.wiiznokes:setting-sliding-windows:2.0.0")
-                implementation("com.example:setting-sliding-windows-jvm:2.0.0")
+                implementation("com.github.wiiznokes:setting-sliding-windows:2.0.1")
+                //implementation("com.example:setting-sliding-windows-jvm:2.0.1")
             }
         }
         val jvmTest by getting {
