+++ FanControl/settings.gradle.kts
@@ -1,11 +1,11 @@
-// Copyright 2000-2022 JetBrains s.r.o. and contributors. Use of this source code is governed by the Apache 2.0 license.
 pluginManagement {
     repositories {
         google()
         gradlePluginPortal()
         mavenCentral()
         maven("https://maven.pkg.jetbrains.space/public/p/compose/dev")
-        mavenLocal()
+        maven("https://jitpack.io")
+        //mavenLocal()
     }
 
     plugins {
