+++ FanControl/src/jvmMain/kotlin/settings/Settings.kt
@@ -7,8 +7,8 @@ import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.update
 import org.json.JSONObject
 import org.json.JSONTokener
-import ui.screen.drawer.secondView.Languages
-import ui.screen.drawer.secondView.Themes
+import ui.screen.drawer.settings.Languages
+import ui.screen.drawer.settings.Themes
 import utils.getJsonValue
 import utils.setJsonValue
 import java.io.File
