+++ FanControl/src/jvmMain/kotlin/ui/screen/drawer/DrawerVM.kt
@@ -5,8 +5,8 @@ import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.update
 import settings.Settings
 import settings.SettingsModel
-import ui.screen.drawer.secondView.Languages
-import ui.screen.drawer.secondView.Themes
+import ui.screen.drawer.settings.Languages
+import ui.screen.drawer.settings.Themes
 
 
 class DrawerVM(
