+++ FanControl/src/jvmMain/kotlin/ui/screen/drawer/drawer.kt
@@ -5,12 +5,15 @@ import androidx.compose.foundation.layout.*
 import androidx.compose.material3.*
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
 import com.example.settingSlidingWindows.Setting
 import com.example.settingSlidingWindows.SettingDefaults
 import com.example.settingSlidingWindows.rememberSettingState
 import kotlinx.coroutines.CoroutineScope
-import ui.screen.drawer.firstView.headerSetting
-import ui.screen.drawer.secondView.*
+import kotlinx.coroutines.launch
+import ui.component.managerText
+import ui.screen.drawer.settings.*
 import ui.utils.Resources
 
 private val viewModel = DrawerVM()
@@ -33,7 +36,7 @@ fun drawer(
         )
     ) {
         header {
-            headerSetting(
+            managerHeader(
                 drawerState = drawerState,
                 scope = scope
             )
@@ -69,5 +72,40 @@ fun drawer(
         group(text = Resources.getString("settings/trans/other"))
         info()
         help()
+    }
+}
+
+
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+private fun managerHeader(
+    drawerState: DrawerState,
+    scope: CoroutineScope,
+) {
+    Column {
+        Row(
+            modifier = Modifier
+                .fillMaxWidth(),
+            horizontalArrangement = Arrangement.SpaceBetween,
+        ) {
+            managerText(
+                modifier = Modifier
+                    .padding(start = 25.dp, top = 40.dp, bottom = 50.dp),
+                text = Resources.getString("title/setting"),
+                color = MaterialTheme.colorScheme.inverseOnSurface,
+                style = MaterialTheme.typography.titleLarge
+            )
+
+            IconButton(
+                onClick = { scope.launch { drawerState.close() } }
+            ) {
+                Icon(
+                    painter = Resources.getIcon("arrow/arrow_back48"),
+                    contentDescription = Resources.getString("ct/close_drawer"),
+                    tint = MaterialTheme.colorScheme.inverseOnSurface
+                )
+            }
+        }
     }
 }
