+++ FanControl/src/jvmMain/kotlin/ui/screen/drawer/settings/info.kt
@@ -1,4 +1,4 @@
-package ui.screen.drawer.secondView
+package ui.screen.drawer.settings
 
 import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
@@ -6,17 +6,18 @@ import com.example.settingSlidingWindows.SettingScope
 import ui.utils.Resources
 
 
-fun SettingScope.help() {
+fun SettingScope.info() {
     item(
-        title = Resources.getString("settings/help"),
-        subTitle = Resources.getString("settings/help_sub_title"),
+        title = Resources.getString("settings/info"),
+        subTitle = Resources.getString("settings/info_sub_title"),
         icon = {
             Icon(
-                painter = Resources.getIcon("settings/help40"),
+                painter = Resources.getIcon("settings/info40"),
                 tint = MaterialTheme.colorScheme.inverseOnSurface,
                 contentDescription = null
             )
-        }
+        },
+        showTopLine = true
     ) {
         Header(null, null)
     }
