+++ FanControl/src/jvmMain/kotlin/ui/screen/drawer/settings/language.kt
@@ -1,4 +1,4 @@
-package ui.screen.drawer.secondView
+package ui.screen.drawer.settings
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.*
