+++ FanControl/src/jvmMain/kotlin/ui/screen/drawer/settings/launchAndExit.kt
@@ -1,8 +1,9 @@
-package ui.screen.drawer.secondView
+package ui.screen.drawer.settings
 
 import androidx.compose.material3.Checkbox
 import androidx.compose.material3.ExperimentalMaterial3Api
 import com.example.settingSlidingWindows.SettingScope
+import ui.theme.Typography
 import ui.utils.Resources
 
 @OptIn(ExperimentalMaterial3Api::class)
@@ -12,6 +13,7 @@ fun SettingScope.exitOnClose(
 ) {
     item(
         title = Resources.getString("settings/exit_on_close"),
+        titleStyle = Typography.titleSmall,
         advanceIconButton = {
             Checkbox(
                 checked = exitOnClose,
@@ -29,6 +31,7 @@ fun SettingScope.launchAtStartUp(
 ) {
     item(
         title = Resources.getString("settings/launch_at_startup"),
+        titleStyle = Typography.titleSmall,
         advanceIconButton = {
             Checkbox(
                 checked = launchAtStartUp,
