+++ FanControl/src/jvmMain/kotlin/ui/theme/Theme.kt
@@ -6,7 +6,7 @@ import androidx.compose.material3.darkColorScheme
 import androidx.compose.material3.lightColorScheme
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.graphics.Color
-import ui.screen.drawer.secondView.Themes
+import ui.screen.drawer.settings.Themes
 
 private val darkColorScheme = darkColorScheme(
